Organize components by category
