+++ components/app/App.tsx
@@ -10,20 +10,20 @@ import * as React from 'react';
 import { useGameLogic } from '../../hooks/useGameLogic';
 import SceneDisplay from '../SceneDisplay';
 import ActionOptions from '../ActionOptions';
-import InventoryDisplay from '../InventoryDisplay';
-import LocationItemsDisplay from '../LocationItemsDisplay';
+import InventoryDisplay from '../inventory/InventoryDisplay';
+import LocationItemsDisplay from '../inventory/LocationItemsDisplay';
 import LoadingSpinner from '../LoadingSpinner';
 import ErrorDisplay from '../ErrorDisplay';
 import QuestInfoBox from '../QuestInfoBox';
 import MainToolbar from '../MainToolbar';
 import ModelUsageIndicators from '../ModelUsageIndicators';
-import TitleMenu from '../TitleMenu';
+import TitleMenu from '../modals/TitleMenu';
 import DialogueDisplay from '../DialogueDisplay';
-import DebugView from '../DebugView';
-import ItemChangeAnimator from '../ItemChangeAnimator';
-import CustomGameSetupScreen from '../CustomGameSetupScreen';
-import SettingsDisplay from '../SettingsDisplay';
-import InfoDisplay from '../InfoDisplay';
+import DebugView from '../debug/DebugView';
+import ItemChangeAnimator from '../inventory/ItemChangeAnimator';
+import CustomGameSetupScreen from '../modals/CustomGameSetupScreen';
+import SettingsDisplay from '../modals/SettingsDisplay';
+import InfoDisplay from '../modals/InfoDisplay';
 import Footer from './Footer';
 import AppModals from './AppModals';
 import AppHeader from './AppHeader';
