+++ components/modals/CustomGameSetupScreen.tsx
@@ -5,10 +5,10 @@
  */
 import { useCallback } from 'react';
 
-import { AdventureTheme, ThemePackName } from '../types';
-import { THEME_PACKS } from '../themes'; // To get pack names and structure
-import Button from './elements/Button';
-import { Icon } from './elements/icons';
+import { AdventureTheme, ThemePackName } from '../../types';
+import { THEME_PACKS } from '../../themes';
+import Button from '../elements/Button';
+import { Icon } from '../elements/icons';
 
 interface CustomGameSetupScreenProps {
   readonly isVisible: boolean;
